Adjustments after rebase
